Fix additional counter problem + fix pathconf() error bug
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@695 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/findlib/find.c
@@ -29,8 +29,8 @@
 #include "find.h"
 
 
-size_t name_max;	       /* filename max length */
-size_t path_max;	       /* path name max length */
+int32_t name_max;	       /* filename max length */
+int32_t path_max;	       /* path name max length */
 
 
 /* ****FIXME**** debug until stable */
